AIAsistant  - Medical Kit/requirements.txt
opencv-python==4.8.1
numpy==1.24.3
Pillow==10.0.0
pygame==2.5.2
python-omxplayer-wrapper==0.3.1
picamera2==0.3.12
psutil==5.9.5 